[Refactor] Delivery 도메인 분리 (#14)
