+++ backend/src/main/java/org/example/backend/domain/company/model/entity/Company.java
@@ -5,10 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.example.backend.domain.user.model.entity.Delivery;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 @Entity
 @Getter
