+++ backend/src/main/java/org/example/backend/domain/delivery/repository/DeliveryRepository.java
@@ -1,6 +1,6 @@
-package org.example.backend.domain.user.repository;
+package org.example.backend.domain.delivery.repository;
 
-import org.example.backend.domain.user.model.entity.Delivery;
+import org.example.backend.domain.delivery.model.entity.Delivery;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
