+++ backend/src/main/java/org/example/backend/domain/user/model/entity/User.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.backend.domain.delivery.model.entity.Delivery;
 import org.example.backend.domain.user.model.dto.UserDto;
 
 import java.time.LocalDateTime;
